requirements.txt
Django
djangorestframework
psycopg2==2.8.6
celery[redis]==5.2.7
flower==1.2.0
celery_singleton==0.3.1
django-redis==5.2.0
python-dotenv
djangorestframework-simplejwt
djoser==2.2.3
django-allauth
django-filter


